requirements.txt
Django==4.2.3
python-telegram-bot==20.4
gunicorn==21.2.0
django-heroku==0.3.1
pre-commit==3.3.3
django_extensions==3.2.3
djangorestframework==3.14.0
dj-database-url==2.0.0
django-constance[database]==2.9.1
django-tinymce==3.6.1
